requirements.txt
fastapi
uvicorn[standard]
torch>=2.4.0
torchvision>=0.19.0
Pillow
psycopg2-binary
python-dotenv
python-multipart
cloudinary
opencv-python-headless
matplotlib
numpy
scipy
timm>=1.0.0
